fix: moving chatbot to app component to prevent unnecessary rerenders
